Refactor, style, and extend user action tests
